# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Africa Tourism Platform API
  version: 1.0.0
  description: API documentation for the Africa Tourism Platform backend.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: StrongPass123
              first_name: Test
              last_name: User
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                message: User registered successfully
                user_id: 00000000-0000-0000-0000-000000000001
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid request
  /api/v1/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: StrongPass123
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                user:
                  id: 00000000-0000-0000-0000-000000000001
                  email: user@example.com
        '401':
          description: Unauthorized
  /api/v1/upload/image:
    post:
      summary: 🔒 [INTERNAL] Upload an image (Backend Test Only)
      description: |
        **Note to Frontend:** Do not use this endpoint directly for Packages, Posts, or Profile updates. 
        Instead, use the multipart support on the respective `POST /packages`, `POST /posts`, or `PATCH /profile` endpoints.
      tags: [Upload]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: folder
          schema:
            type: string
          description: Target folder (e.g., packages, profiles, posts)
          example: packages
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (max 10MB)
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: url
              example:
                url: https://res.cloudinary.com/demo/image/upload/v12345/africa-tourism/packages/sample.jpg
        '400':
          description: Invalid file or too large
  /api/v1/upload/video:
    post:
      summary: 🔒 [INTERNAL] Upload a video (Backend Test Only)
      description: |
        **Note to Frontend:** Do not use this endpoint directly for Community posts. 
        Instead, use the multipart support on the `POST /posts` endpoint.
      tags: [Upload]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: folder
          schema:
            type: string
          description: Target folder (e.g., posts)
          example: posts
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Video file (max 100MB)
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: url
              example:
                url: https://res.cloudinary.com/demo/video/upload/v12345/africa-tourism/posts/sample.mp4
        '400':
          description: Invalid file or too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid email or password
  /api/v1/packages:
    get:
      summary: Get packages feed
      tags: [Packages]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [public, private, archived, all]
        - in: query
          name: sort_by
          schema:
            type: string
            enum: [rating_avg, price, verified, views]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackagesFeedResponse'
              example:
                data:
                  - package_id: 00000000-0000-0000-0000-000000000500
                    creator_id: 00000000-0000-0000-0000-000000000001
                    itinerary_id: 00000000-0000-0000-0000-000000000700
                    title: Safari Adventure
                    summary: A 3-day guided safari
                    price: 299.99
                    status: public
                    rating_avg: 4.8
                    reviews_count: 10
                    views_count: 120
                    country: Kenya
                    location: Nairobi
                    currency: USD
                    image_url: https://example.com/safari.jpg
                    duration_days: 3
                    category: adventure
                    group_size: small
                    is_public: true
                    created_at: "2026-03-14T10:00:00Z"
                meta:
                  page: 1
                  page_size: 20
                  total: 1
                  sort_by: rating_avg
                  order: desc
    post:
      summary: Create package
      tags: [Packages]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageCreateRequest'
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/PackageCreateRequest'
                - type: object
                  properties:
                    image:
                      type: string
                      format: binary
                      description: Optional package thumbnail image
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
              example:
                package_id: 00000000-0000-0000-0000-000000000500
                creator_id: 00000000-0000-0000-0000-000000000001
                itinerary_id: 00000000-0000-0000-0000-000000000700
                title: Safari Adventure
                summary: A 3-day guided safari
                price: 299.99
                status: private
                rating_avg: 0
                reviews_count: 0
                views_count: 0
                country: Kenya
                location: Nairobi
                currency: USD
                image_url: https://example.com/safari.jpg
                duration_days: 3
                category: adventure
                is_public: false
                created_at: "2026-03-14T10:00:00Z"
  /api/v1/packages/{id}:
    get:
      summary: Get package by ID
      tags: [Packages]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
              example:
                package_id: 00000000-0000-0000-0000-000000000500
                creator_id: 00000000-0000-0000-0000-000000000001
                itinerary_id: 00000000-0000-0000-0000-000000000700
                title: Safari Adventure
                summary: A 3-day guided safari
                price: 299.99
                status: public
                rating_avg: 4.8
                reviews_count: 10
                views_count: 121
                country: Kenya
                location: Nairobi
                currency: USD
                image_url: https://example.com/safari.jpg
                duration_days: 3
                category: adventure
                is_public: true
                created_at: "2026-03-14T10:00:00Z"
    patch:
      summary: Update package
      tags: [Packages]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageUpdateRequest'
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/PackageUpdateRequest'
                - type: object
                  properties:
                    image:
                      type: string
                      format: binary
                      description: Optional new package thumbnail image
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
              example:
                package_id: 00000000-0000-0000-0000-000000000500
                creator_id: 00000000-0000-0000-0000-000000000001
                itinerary_id: 00000000-0000-0000-0000-000000000700
                title: Updated Package Title
                summary: Updated summary
                price: 149.5
                status: private
                rating_avg: 0
                reviews_count: 0
                views_count: 0
                country: Kenya
                location: Nairobi
                currency: USD
                image_url: https://example.com/safari.jpg
                duration_days: 3
                category: adventure
                is_public: false
                created_at: "2026-03-14T10:00:00Z"
    delete:
      summary: Archive package
      tags: [Packages]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Package archived successfully
  /api/v1/packages/{id}/publish:
    post:
      summary: Publish package
      tags: [Packages]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Package public successfully
  /api/v1/packages/{id}/status:
    patch:
      summary: Update package status
      tags: [Packages]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageStatusRequest'
            example:
              status: public
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Status updated successfully
  /api/v1/packages/{id}/reviews:
    get:
      summary: Get package reviews
      tags: [Packages]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackageReview'
              example:
                - review_id: 00000000-0000-0000-0000-000000000900
                  package_id: 00000000-0000-0000-0000-000000000500
                  user_id: 00000000-0000-0000-0000-000000000001
                  rating: 5
                  comment: Great trip
                  created_at: "2026-03-14T10:00:00Z"
    post:
      summary: Submit package review
      tags: [Packages]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
            example:
              rating: 5
              comment: Great trip
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Review submitted successfully
  /api/v1/packages/{id}/chat:
    get:
      summary: Get package chat history
      tags: [Chat]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatHistoryResponse'
              example:
                data:
                  - chat_id: 00000000-0000-0000-0000-000000001000
                    package_id: 00000000-0000-0000-0000-000000000500
                    user_id: 00000000-0000-0000-0000-000000000001
                    message: Is transport included?
                    created_at: "2026-03-14T10:00:00Z"
                meta:
                  page: 1
                  page_size: 50
                  total: 1
    post:
      summary: Post chat message
      tags: [Chat]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              message: Is transport included?
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageChat'
              example:
                chat_id: 00000000-0000-0000-0000-000000001000
                package_id: 00000000-0000-0000-0000-000000000500
                user_id: 00000000-0000-0000-0000-000000000001
                message: Is transport included?
                created_at: "2026-03-14T10:00:00Z"
  /api/v1/packages/{id}/ws:
    get:
      summary: WebSocket chat
      description: WebSocket endpoint for package chat. Use a WebSocket client and include Authorization header.
      tags: [Chat]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '101':
          description: Switching protocols
  /api/v1/user/profile:
    get:
      summary: Get user profile
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfileResponse'
              example:
                data:
                  user_id: 00000000-0000-0000-0000-000000000001
                  email: user@example.com
                  first_name: Arsema
                  last_name: Zeweldi
                  country: Ethiopia
                  bio: Passionate traveler and developer.
                  profile_image_url: https://example.com/avatar.jpg
                  created_at: 2026-03-14T10:00:00Z
    patch:
      summary: Update user profile
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/UpdateProfileRequest'
                - type: object
                  properties:
                    avatar:
                      type: string
                      format: binary
                      description: Optional user avatar image
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserProfileResponse'
              example:
                message: Profile updated
                data:
                  user_id: 00000000-0000-0000-0000-000000000001
                  email: user@example.com
                  first_name: Arsema
                  last_name: Zeweldi
                  country: Ethiopia
                  bio: Updated bio.
                  profile_image_url: https://example.com/avatar.jpg
                  created_at: 2026-03-14T10:00:00Z
  /api/v1/user/preferences:
    get:
      summary: Get user preferences
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesResponse'
              example:
                data:
                  preference_id: 00000000-0000-0000-0000-000000000300
                  user_id: 00000000-0000-0000-0000-000000000001
                  preferred_season: summer
                  budget_range: medium
                  preferred_activities: [hiking, food]
                  dietary_restrictions: [vegetarian]
                  preferred_climate: tropical
                  preferred_language: english
                  travel_vibe_interest: adventure
    put:
      summary: Update user preferences
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preferred_season:
                  type: string
                  enum: [spring, summer, autumn, winter, any]
                budget_range:
                  type: string
                  enum: [low, medium, high, luxury]
                preferred_activities:
                  type: array
                  items:
                    type: string
                dietary_restrictions:
                  type: array
                  items:
                    type: string
                preferred_climate:
                  type: string
                  enum: [tropical, desert, coastal, highland, temperate, any]
                preferred_language:
                  type: string
                  enum: [english, french, arabic, portuguese, swahili, any]
                travel_vibe_interest:
                  type: string
                  enum: [adventure, relaxed, foodie, history, party, culture, wildlife, any]
                travel_vibes:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Preferences updated
  /api/v1/planner/generate:
    post:
      summary: Generate itinerary using AI
      tags: [Planner]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePlanRequest'
            example:
              destination: Nairobi, Kenya
              duration_days: 3
              budget_level: medium
              vibe_tags: [culture, food]
              group_size: 2
              climate_pref: any
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratePlanResponse'
              example:
                message: Plan generated
                data:
                  title: Nairobi Highlights
                  currency: USD
                  days:
                    - day_num: 1
                      activities:
                        - name: City tour
                          type: culture
                          description: Explore downtown
                          est_cost: 50
                          geo_lat: -1.286389
                          geo_long: 36.817223
  /api/v1/itineraries:
    get:
      summary: List user itineraries
      tags: [Itineraries]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Itinerary'
    post:
      summary: Create itinerary
      tags: [Itineraries]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItineraryRequest'
            example:
              title: Test Itinerary
              description: Test itinerary description
              days_count: 3
              nights_count: 2
              start_date: "2026-03-20"
              end_date: "2026-03-23"
              items:
                - day_number: 1
                  title: Arrival and check-in
                  description: Check in and rest
                  start_time: "09:00"
                  end_time: "12:00"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'
              example:
                itinerary_id: 00000000-0000-0000-0000-000000000700
                user_id: 00000000-0000-0000-0000-000000000001
                title: Test Itinerary
                description: Test itinerary description
                days_count: 3
                nights_count: 2
                items: []
  /api/v1/itineraries/{id}:
    get:
      summary: Get itinerary by ID
      tags: [Itineraries]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'
              example:
                itinerary_id: 00000000-0000-0000-0000-000000000700
                user_id: 00000000-0000-0000-0000-000000000001
                title: Test Itinerary
                description: Test itinerary description
                days_count: 3
                nights_count: 2
                items:
                  - item_id: 00000000-0000-0000-0000-000000000701
                    itinerary_id: 00000000-0000-0000-0000-000000000700
                    day_number: 1
                    title: Arrival
    delete:
      summary: Delete itinerary
      tags: [Itineraries]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Itinerary deleted successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: StrongPass123
        first_name:
          type: string
        last_name:
          type: string
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        user_id:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: StrongPass123
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
    Package:
      type: object
      properties:
        package_id:
          type: string
          format: uuid
        creator_id:
          type: string
          format: uuid
        itinerary_id:
          type: string
          format: uuid
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [private, public, archived]
        rating_avg:
          type: number
        reviews_count:
          type: integer
        views_count:
          type: integer
        country:
          type: string
        location:
          type: string
        currency:
          type: string
        image_url:
          type: string
        duration_days:
          type: integer
        category:
          type: string
        group_size:
          type: string
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PackageCreateRequest:
      type: object
      required: [itinerary_id, title]
      properties:
        itinerary_id:
          type: string
          format: uuid
        title:
          type: string
        summary:
          type: string
        price:
          type: number
        country:
          type: string
        location:
          type: string
        currency:
          type: string
        image_url:
          type: string
        duration_days:
          type: integer
        category:
          type: string
        group_size:
          type: string
        is_public:
          type: boolean
...
    PackagesFeedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        meta:
          type: object
          properties:
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
...